ansible: add disable option for each callback (#1705)

+++ plugins/tasks/ansible/src/main/java/com/walmartlabs/concord/plugins/ansible/OutVarsProcessor.java
@@ -60,7 +60,7 @@ public class OutVarsProcessor {
     }
 
     public void postProcess() throws IOException {
-        if (outVarsFile != null) {
+        if (outVarsFile != null && Files.exists(outVarsFile)) {
             Files.delete(outVarsFile);
         }
     }
